Revert groupy move - not the same fn
