+++ packages/dashboard/src/app/api/useCategoriesApi.js
@@ -23,11 +23,12 @@ import {
   useMemo,
   useState,
 } from '@web-stories-wp/react';
-import { addQueryArgs, groupBy } from '@web-stories-wp/design-system';
+import { addQueryArgs } from '@web-stories-wp/design-system';
 
 /**
  * Internal dependencies
  */
+import groupBy from '../../utils/groupBy';
 import fetchAllFromTotalPages from './fetchAllFromPages';
 
 export default function useCategoriesApi(dataAdapter, { categoryApi }) {
