+++ packages/design-system/src/utils/index.js
@@ -19,6 +19,5 @@ export { noop } from './noop';
 export { default as addQueryArgs } from './addQueryArgs';
 export { default as labelAccessibilityValidator } from './labelAccessibilityValidator';
 export { default as useLiveRegion } from './useLiveRegion';
-export { default as groupBy } from './groupBy';
 export * from './constants';
 export * from './directions';
